V1 terminé avec securisation de l'espace admin

+++ admin/crud/suppression_carousel_header.php
@@ -1,13 +1,26 @@
 <?php 
+  session_start();
+?>
+<?php 
 
 include("../../bdd.php");
 
-$page = $_GET['page'];
-$table = $_GET['table'];
+$connexion = $bdd->prepare("SELECT pseudo, mot_de_passe FROM admin WHERE pseudo=?");
+$connexion->execute(array($_SESSION['pseudo']));
+$connexion_recu = $connexion->fetch();
+if ($_SESSION['pseudo'] == $connexion_recu['pseudo'] && password_verify($_SESSION['mot_de_passe'], $connexion_recu['mot_de_passe'])){
 
-$suppression = $bdd->prepare("DELETE FROM $table WHERE id=?");
-$suppression->execute(array($_GET['id']));
+    $page = $_GET['page'];
+    $table = $_GET['table'];
 
-header("location:$page");
+    $suppression = $bdd->prepare("DELETE FROM $table WHERE id=?");
+    $suppression->execute(array($_GET['id']));
+
+    header("location:$page");
+
+}
+else {
+  header("location:../erreur_connexion.php");
+}
 
 ?>
